FEAT: divided roomWS and bookingWS from hotelMTA, still not working makeBooking
