+++ src/main/java/common/consts/WebMethodConsts.java
@@ -9,5 +9,5 @@ public class WebMethodConsts {
     public static final String OP_USER_REGISTRATION = "registerUser";
     public static final String OP_USER_LOGIN = "loginUser";
     public static final String OP_SEARCH_ROOM = "searchRoom";
-    public static final String OP_MAKE_ROOM_RESERVATION = "makeHotelBooking";
+    public static final String OP_MAKE_HOTEL_BOOKING = "makeHotelBooking";
 }
